# Speedup Xcbeautify by applying frequency-based pattern matching.