--- a/Sources/xcbeautify/Xcbeautify.swift
+++ b/Sources/xcbeautify/Xcbeautify.swift
@@ -27,7 +27,6 @@ struct Xcbeautify: ParsableCommand {
     var reportPath = "build/reports"
 
     func run() throws {
-        let parser = Parser()
         let output = OutputHandler(quiet: quiet, quieter: quieter, isCI: isCi, { print($0) })
         let junitReporter = JunitReporter()
 
@@ -40,11 +39,11 @@ struct Xcbeautify: ParsableCommand {
             }
             return line
         }
+        
+        let parser = Parser(colored: !disableColoredOutput, additionalLines: { readLine() })
 
         while let line = readLine() {
-            guard let formatted = parser.parse(line: line,
-                                               colored: !disableColoredOutput,
-                                               additionalLines: { readLine() }) else { continue }
+            guard let formatted = parser.parse(line: line) else { continue }
             output.write(parser.outputType, formatted)
         }