The current implementation of xcbeautify performs regex checking for each line of the xcodebuild output.
Due to the big amount of possible regexes, each line on average is checked for about `N/2` times. which is a lot

Suggested implementation holds an array of 'inner parser' and try to reuse 'line parser/matchers' using frequency based array (basically, last matcher regex will be checked first).
Due to the non-random structure of xcodebuil output, this approach allows to decrease `xcbeautify` running time from 160 s to 9 seconds when used on the xcodebuild output of size 240Mb

| Implementation | time | Log Size |
|-|-|-|
| xcbeautify-main | 209s | 256M |
| xcbeautify-speedup | 8s | 256M |
| xcpretty | 23s | 256M |

Testing approach: 
`time cat LOGFILE | xcbeauitify`

## Possible 'bugs'
If there are multiple regexes which can match same line, this approach could return different results, depending on the order of the commands, appeared in log

## Xcbeautify Speed
It seems that xcbeautify is waaaay slower than xcpretty. This PR will actually make xcbeautify faster